Refactored api.js and split the file to separate api files, depending on the reducers that use those files
